requirements.txt
requests
python-dotenv
sqlalchemy
playwright
beautifulsoup4
openai
